Public sealed scope state
It is currently impossible to check the status of the creation or deletion of a scope. Therefore, we export the ScopeState, which reflects the current state of the scope.

---

I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=ru

---

Pull Request resolved: https://github.com/yandex/yx_scope/pull/22
commit_hash:3a3cba954ba16d49da888a17a19cbff83a596326